moved breakpointMaps into bpMapConstants.js
