+++ client/src/components/ProductCardItem/ProductCardItem.js
@@ -24,16 +24,17 @@ const ProductCardItem = ({
   name,
   rating,
   _id,
-  isTranding,
+  isTrending,
   isSupreme,
   isHealthy,
   itemNo,
 }) => {
-  const breackPoint = useBreakpoint();
+  const breakPoint = useBreakpoint();
   const products = useSelector((state) => state.products.products);
   const isUserAuthorized = useSelector((state) => state.authorization.isUserAuthorized);
   const dispatch = useDispatch();
   const randomNum = Math.floor(Math.random() * (59 - 29 + 1)) + 29;
+
   const handleOpenModalLogin = () => {
     dispatch(openModal());
     dispatch(setContent(<LoginForm />));
@@ -81,7 +82,7 @@ const ProductCardItem = ({
           <Box sx={stylesMiniTextWrapper}>
             <ColorChips
               isHealthy={isHealthy}
-              isTranding={isTranding}
+              isTrending={isTrending}
               isSupreme={isSupreme}
               customStyles={stylesMiniText}
             />
@@ -106,7 +107,7 @@ const ProductCardItem = ({
         </Typography>
       </Box>
       <CardActions onClick={handleAddToCart} sx={stylesButton}>
-        {breackPoint !== 'mobile' ? (<b>ADD</b>) : null}
+        {breakPoint !== 'mobile' ? (<b>ADD</b>) : null}
         <ShoppingCartCheckoutIcon />
       </CardActions>
     </>
@@ -120,7 +121,7 @@ ProductCardItem.propTypes = {
   rating: PropTypes.number,
   _id: PropTypes.string,
   isHealthy: PropTypes.bool,
-  isTranding: PropTypes.bool,
+  isTrending: PropTypes.bool,
   isSupreme: PropTypes.bool,
   itemNo: PropTypes.string,
 };
@@ -132,7 +133,7 @@ ProductCardItem.defaultProps = {
   rating: '',
   _id: '',
   isHealthy: null,
-  isTranding: null,
+  isTrending: null,
   isSupreme: null,
   itemNo: '',
 };
