+++ client/src/constants/bpMapConstants.js
@@ -1,0 +1,58 @@
+export const topProductsQtyMap = {
+  mobile: 4,
+  tablet: 4,
+  lgTablet: 3,
+  desktop: 5,
+};
+
+export const topPartnersQtyMap = {
+  mobile: 3,
+  tablet: 3,
+  lgTablet: 2,
+  desktop: 3,
+};
+
+export const productsPerPageMap = {
+  mobile: 10,
+  tablet: 10,
+  lgTablet: 9,
+  desktop: 15,
+};
+
+export const groupOfStarsSizesMap = {
+  mobile: {
+    width: '188px',
+    height: '252px',
+  },
+  tablet: {
+    width: '188px',
+    height: '252px',
+  },
+  lgTablet: {
+    width: '188px',
+    height: '252px',
+  },
+  desktop: {
+    width: '240px',
+    height: '360px',
+  },
+};
+
+export const googleSvgSizesMap = {
+  mobile: {
+    width: 28,
+    height: 27,
+  },
+  tablet: {
+    width: 31,
+    height: 30,
+  },
+  lgTablet: {
+    width: 33,
+    height: 32,
+  },
+  desktop: {
+    width: 37,
+    height: 36,
+  },
+};
